Gestion des likes

+++ p6-back/controllers/sauces.js
@@ -4,11 +4,13 @@ import Sauce from "../models/sauces.js";
 function normalizer(sauce, req){
     return {
         ...sauce, 
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${sauce.imageName}`
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${sauce.imageName}`,
+        likes: sauce.usersLiked.length,
+        dislikes: sauce.usersDisliked.length
     }
 }
 
-export async function listeSauces(req, res, next){
+export async function listeSauces(req, res){
     // res.status(200).json([])
     try {
         const sauces = await Sauce.find();
@@ -21,7 +23,6 @@ export async function listeSauces(req, res, next){
 }
 
 export async function createSauce(req, res){
-    // console.log(req.file);
     try {
         const body = JSON.parse(req.body.sauce);
         const sauce = new Sauce( {
@@ -58,10 +59,19 @@ export async function displaySauce(req, res){
 
 export async function deleteSauce(req, res){
     try {
+        const userId = req.auth.userId;
+        const sauce = await Sauce.findOne(
+            {_id: req.params.id}
+        ); 
+
+        if(userId !== sauce.userId){
+            return res.status(401).json({ message : "Minterdiction de supprimer la sauce !" });
+        }
+
         await Sauce.deleteOne({
             _id: req.params.id
         });
-        
+
         res.status(201).json({ message: 'Sauce supprimée !' });
     } catch (error) {
         console.error(error);
@@ -71,6 +81,15 @@ export async function deleteSauce(req, res){
 
 export async function modifierSauce(req, res){
     try {
+        const userId = req.auth.userId;
+        const sauce = await Sauce.findOne(
+            {_id: req.params.id}
+        ); 
+
+        if(userId !== sauce.userId){
+            return res.status(401).json({ message : "Interdiction de maudifier la sauce !" });
+        }
+
         await Sauce.updateOne(
             { _id: req.params.id },
             { ...req.body, _id: req.params.id }
@@ -83,15 +102,25 @@ export async function modifierSauce(req, res){
     }
 }
 
-
 export async function likeSauce(req, res){
     try {
         const sauce = await Sauce.findOne(
             {_id: req.params.id}
         ); 
         const userId = req.auth.userId;
-        sauce.usersLiked.push(userId);
+        const like = req.body.like;
+
+        sauce.usersLiked = sauce.usersLiked.filter(userLike => userLike !== userId);
+        sauce.usersDisliked = sauce.usersDisliked.filter(userDislike => userDislike !== userId);
+
+        if(like === 1){
+            sauce.usersLiked.push(userId);
+        }else if(like === -1){
+            sauce.usersDisliked.push(userId);
+        }
+
         await sauce.save();
+
         res.status(200).json({ message: 'Sauce Likée !' });
     } catch (error) {
         console.error(error);
